refactor: extract session continuity from AgentWorkspace
